Fixed league leaderboard routing
