+++ src/views/ApplicationViews.jsx
@@ -1,17 +1,9 @@
-import { useEffect, useState } from "react"
 import { AdministratorViews } from "./AdministratorViews.jsx"
 import { CompetitorViews } from "./CompetitorViews.jsx"
 
 
-export const ApplicationViews = () => {
-  const [currentUser, setCurrentUser] = useState({})
+export const ApplicationViews = ({ currentUser }) => {
 
-  useEffect(() => {
-    const localVertiUser = localStorage.getItem("verti_user")
-    const vertiUserObject = JSON.parse(localVertiUser)
 
-    setCurrentUser(vertiUserObject)
-  }, [])
-
-  return currentUser.isStaff ? <AdministratorViews currentUser={currentUser} /> : <CompetitorViews currentUser={currentUser}/>
+  return currentUser?.isStaff ? <AdministratorViews currentUser={currentUser} /> : <CompetitorViews currentUser={currentUser}/>
 }
