+++ src/views/CompetitorViews.jsx
@@ -2,6 +2,7 @@ import { Outlet, Route, Routes } from "react-router-dom"
 import { Dashboard } from "../components/dashboard/Dashboard.jsx"
 import { CompetitorNavBar } from "../components/navbar/CompetitorNavBar.jsx"
 import { LeagueLeaderboard } from "../components/leaderboards/LeagueLeaderboard.jsx"
+import { CompetitorValidate } from "../components/validate/CompetitorValidate.jsx"
 
 export const CompetitorViews = ({ currentUser }) => {
     return (
@@ -16,6 +17,9 @@ export const CompetitorViews = ({ currentUser }) => {
                 }
             >
                 <Route index element={<Dashboard />} />
+                <Route path="validate">
+                    <Route index element={<CompetitorValidate currentUser={currentUser} />} />
+                </Route>
                 <Route path="leagueLeaderboard">
                     <Route index element={<LeagueLeaderboard />} />
                 </Route>
